requirements.txt
gunicorn==22.0.0
blinker==1.9.0
Bootstrap-Flask==2.4.1
click==8.1.7
colorama==0.4.6
dominate==2.9.1
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
